# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API pública de Asta",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "https://{servidor}"
    }
  ],
...
  "security": [
    {
      "ApiKey": []
    }
  ],
  "paths": {
    "/api/v1/public/invoices": {
      "get": {
        "operationId": "listarFacturas",
        "summary": "Listar tus facturas",
        "description": "Facturas contabilizadas de tu empresa, de la más reciente a la más antigua. Solo facturas: las notas de crédito no aparecen.\n\nPermiso necesario: `INVOICES_READ`.",
        "tags": [
          "Facturas"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "desde",
            "in": "query",
            "required": false,
            "description": "Solo facturas emitidas desde esta fecha, inclusive (`YYYY-MM-DD`).",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "required": false,
            "description": "Solo facturas emitidas hasta esta fecha, inclusive (`YYYY-MM-DD`).",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "estadoPago",
            "in": "query",
            "required": false,
            "description": "Solo facturas en este estado de pago.",
            "schema": {
              "type": "string",
              "enum": [
                "not_paid",
                "in_payment",
                "paid",
                "partial",
                "reversed",
                "invoicing_legacy"
              ]
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "description": "Página, empezando en 1.",
            "schema": {
              "default": 1,
              "type": "integer",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "porPagina",
            "in": "query",
            "required": false,
            "description": "Resultados por página, hasta 100.",
            "schema": {
              "default": 50,
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Una página de facturas.",
            "headers": {
              "RateLimit-Limit": {
                "description": "Peticiones permitidas en la ventana.",
                "schema": {
                  "type": "integer"
                }
              },
              "RateLimit-Remaining": {
                "description": "Peticiones que quedan en la ventana.",
                "schema": {
                  "type": "integer"
                }
              },
              "RateLimit-Reset": {
                "description": "Segundos hasta que se reinicia la ventana.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "exclusiveMinimum": 0,
                            "description": "Identificador de la factura. Es el que se usa en `/invoices/{id}`."
                          },
                          "folio": {
                            "type": "string",
                            "description": "Número de la factura tal como aparece impreso."
                          },
                          "fecha": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date"
                              },
                              {
                                "type": "null"
                              }
                            ],
                            "description": "Fecha de emisión (`YYYY-MM-DD`)."
                          },
                          "vencimiento": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date"
                              },
                              {
                                "type": "null"
                              }
                            ],
                            "description": "Fecha de vencimiento (`YYYY-MM-DD`). `null` si no tiene."
                          },
                          "estadoPago": {
                            "type": "string",
                            "enum": [
                              "not_paid",
                              "in_payment",
                              "paid",
                              "partial",
                              "reversed",
                              "invoicing_legacy",
                              "desconocido"
                            ],
                            "description": "Estado de pago: `not_paid` sin pagar, `partial` pagada en parte, `in_payment` pago registrado pendiente de conciliar, `paid` pagada, `reversed` anulada con nota de crédito, `invoicing_legacy` importada de un sistema anterior, `desconocido` sin estado calculado. Puede recibir valores nuevos: trata cualquier otro como `desconocido`."
                          },
                          "total": {
                            "type": "number",
                            "description": "Importe total, en la moneda de la compañía que emite la factura."
                          },
                          "saldo": {
                            "type": "number",
                            "description": "Importe pendiente de pago, en la misma moneda que `total`."
                          }
                        },
                        "required": [
                          "id",
                          "folio",
                          "fecha",
                          "vencimiento",
                          "estadoPago",
                          "total",
                          "saldo"
                        ]
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagina": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "description": "Página devuelta."
                        },
                        "porPagina": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "description": "Tamaño de página aplicado."
                        },
                        "total": {
                          "type": "integer",
                          "minimum": 0,
                          "description": "Total de resultados con estos filtros, en todas las páginas."
                        }
                      },
                      "required": [
                        "pagina",
                        "porPagina",
                        "total"
                      ]
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ]
                },
                "example": {
                  "data": [
                    {
                      "id": 1001,
                      "folio": "F-0001",
                      "fecha": "2026-09-14",
                      "vencimiento": "2026-10-14",
                      "estadoPago": "not_paid",
                      "total": 1250.4,
                      "saldo": 1250.4
                    }
                  ],
                  "meta": {
                    "pagina": 1,
                    "porPagina": 50,
                    "total": 1
                  }
                }
              }
            }
          },
          "400": {
            "description": "Códigos posibles: `INVALID_QUERY`, `PARTNER_ID_NOT_ALLOWED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Códigos posibles: `MISSING_API_KEY`, `INVALID_API_KEY`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Códigos posibles: `INSUFFICIENT_SCOPE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Códigos posibles: `RATE_LIMITED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Segundos que hay que esperar.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Códigos posibles: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Códigos posibles: `ODOO_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "label": "curl",
            "source": "curl \"https://{servidor}/api/v1/public/invoices?estadoPago=not_paid&porPagina=20\" \\\n  -H \"X-API-Key: $ASTA_API_KEY\""
          },
          {
            "lang": "Python",
            "label": "Python",
            "source": "import json, os, urllib.request\n\npeticion = urllib.request.Request(\n    \"https://{servidor}/api/v1/public/invoices?estadoPago=not_paid&porPagina=20\",\n    headers={\"X-API-Key\": os.environ[\"ASTA_API_KEY\"]},\n)\nwith urllib.request.urlopen(peticion) as respuesta:\n    datos = json.load(respuesta)\nprint(datos)"
          }
        ]
      }
    },
    "/api/v1/public/invoices/{id}": {
      "get": {
        "operationId": "verFactura",
        "summary": "Ver una factura",
        "description": "Una factura de tu empresa. Una factura que no existe y una de otra empresa dan la misma respuesta 404.\n\nPermiso necesario: `INVOICES_READ`.",
        "tags": [
          "Facturas"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Identificador de la factura, el `id` del listado.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La factura.",
            "headers": {
              "RateLimit-Limit": {
                "description": "Peticiones permitidas en la ventana.",
                "schema": {
                  "type": "integer"
                }
              },
              "RateLimit-Remaining": {
                "description": "Peticiones que quedan en la ventana.",
                "schema": {
                  "type": "integer"
                }
              },
              "RateLimit-Reset": {
                "description": "Segundos hasta que se reinicia la ventana.",
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "exclusiveMinimum": 0,
                          "description": "Identificador de la factura. Es el que se usa en `/invoices/{id}`."
                        },
                        "folio": {
                          "type": "string",
                          "description": "Número de la factura tal como aparece impreso."
                        },
                        "fecha": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "Fecha de emisión (`YYYY-MM-DD`)."
                        },
                        "vencimiento": {
                          "anyOf": [
                            {
                              "type": "string",
                              "format": "date"
                            },
                            {
                              "type": "null"
                            }
                          ],
                          "description": "Fecha de vencimiento (`YYYY-MM-DD`). `null` si no tiene."
                        },
                        "estadoPago": {
                          "type": "string",
                          "enum": [
                            "not_paid",
                            "in_payment",
                            "paid",
                            "partial",
                            "reversed",
                            "invoicing_legacy",
                            "desconocido"
                          ],
                          "description": "Estado de pago: `not_paid` sin pagar, `partial` pagada en parte, `in_payment` pago registrado pendiente de conciliar, `paid` pagada, `reversed` anulada con nota de crédito, `invoicing_legacy` importada de un sistema anterior, `desconocido` sin estado calculado. Puede recibir valores nuevos: trata cualquier otro como `desconocido`."
                        },
                        "total": {
                          "type": "number",
                          "description": "Importe total, en la moneda de la compañía que emite la factura."
                        },
                        "saldo": {
                          "type": "number",
                          "description": "Importe pendiente de pago, en la misma moneda que `total`."
                        }
                      },
                      "required": [
                        "id",
                        "folio",
                        "fecha",
                        "vencimiento",
                        "estadoPago",
                        "total",
                        "saldo"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                },
                "example": {
                  "data": {
                    "id": 1001,
                    "folio": "F-0001",
                    "fecha": "2026-09-14",
                    "vencimiento": "2026-10-14",
                    "estadoPago": "paid",
                    "total": 1250.4,
                    "saldo": 0
                  }
                }
              }
            }
          },
          "400": {
            "description": "Códigos posibles: `INVALID_INVOICE_ID`, `PARTNER_ID_NOT_ALLOWED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Códigos posibles: `MISSING_API_KEY`, `INVALID_API_KEY`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Códigos posibles: `INSUFFICIENT_SCOPE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Códigos posibles: `INVOICE_NOT_FOUND`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Códigos posibles: `RATE_LIMITED`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Segundos que hay que esperar.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "Códigos posibles: `INTERNAL_ERROR`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Códigos posibles: `ODOO_UNAVAILABLE`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "label": "curl",
            "source": "curl \"https://{servidor}/api/v1/public/invoices/1001\" \\\n  -H \"X-API-Key: $ASTA_API_KEY\""
          },
          {
            "lang": "Python",
            "label": "Python",
            "source": "import json, os, urllib.request\n\npeticion = urllib.request.Request(\n    \"https://{servidor}/api/v1/public/invoices/1001\",\n    headers={\"X-API-Key\": os.environ[\"ASTA_API_KEY\"]},\n)\nwith urllib.request.urlopen(peticion) as respuesta:\n    datos = json.load(respuesta)\nprint(datos)"
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Tu API key. Se crea en el panel, en «Mis API keys»."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Código del error. Programa contra él, no contra `message`. Pueden aparecer códigos nuevos: trátalos por su status HTTP.",
                "examples": [
                  "MISSING_API_KEY",
                  "INVALID_API_KEY",
                  "INSUFFICIENT_SCOPE",
                  "PARTNER_ID_NOT_ALLOWED",
                  "INVALID_QUERY",
                  "INVALID_INVOICE_ID",
                  "INVOICE_NOT_FOUND",
                  "INVOICE_PDF_NOT_AVAILABLE",
                  "LINK_NOT_VALID",
                  "LINK_EXPIRED",
                  "PRINTER_NOT_FOUND",
                  "INVENTORY_UNAVAILABLE",
                  "RATE_LIMITED",
                  "VALIDATION_ERROR",
                  "ODOO_UNAVAILABLE",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "required": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "docs": {
                "type": "string",
                "format": "uri"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}